requirements.txt
streamlit
Pillow
requests
numpy
torch==1.13.1
ultralytics==8.2.103
opencv-python-headless==4.8.0.74
# prevent normal opencv from being installed
opencv-python==0.0.0
